[scenetalk] Add SceneTalkConnector component to make UI easier to assemble.
