+++ sites/jungle3/src/pages/PackageScene.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, useRef, useCallback } from "react";
+import { useEffect, useState, useCallback } from "react";
 import {
   Box,
   Button,
@@ -21,7 +21,7 @@ import SceneViewer from "@components/BabylonViewer/SceneViewer";
 import SceneControls from "@components/BabylonViewer/SceneControls";
 import GeneratorSelector from "@components/BabylonViewer/GeneratorSelector";
 import { useSceneStore } from "scenetalk";
-import { SceneTalkConnection } from "scenetalk";
+import { SceneTalkConnector } from "scenetalk";
 import { useWindowSize } from "@hooks/useWindowSize";
 import { useNavigate, useParams } from "react-router-dom";
 import { useGetAssetByVersion, useGetJobDefinition } from "@queries/packages";
@@ -34,33 +34,14 @@ export const PackageScene: React.FC = () => {
   const { asset_id, version_id } = useParams();
   const { t } = useTranslation();
   const [isLoading, setIsLoading] = useState(true);
-  const wsServiceRef = useRef<SceneTalkConnection | null>(null);
   const { currentWidth } = useWindowSize();
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [isSidepanelOpen, setIsSidepanelOpen] = useState(true); // New state for panel visibility
 
   // Get state and actions from the store
   const {
-    wsStatus,
-    setWsStatus,
     selectedHdaId,
-    paramValues,
     setParamValues,
-    fileUpload,
-    inputFiles,
-    setMeshData,
-    addStatusLog,
-    flushStatusLog,
-    clearStatusLog,
-    setGenerateTime,
-    requestInFlight,
-    setRequestInFlight,
-    pendingRequest,
-    setPendingRequest,
-    exportFormat,
-    setExportFormat,
-    setLatency,
-    reset
   } = useSceneStore();
 
   const navigate = useNavigate();
@@ -71,80 +52,6 @@ export const PackageScene: React.FC = () => {
   const { data: assetVersion } =
     useGetAssetByVersion(asset_id as string, version_id as string);
 
-  // Initialize WebSocket service
-  useEffect(() => {
-    const sceneTalkUrl = import.meta.env.VITE_SCENE_TALK_URL;
-    const wsService = new SceneTalkConnection(sceneTalkUrl);
-    wsServiceRef.current = wsService;
-    wsService.connect();
-
-    return () => {
-      wsService.disconnect();
-    };
-  }, []);
-
-  // Update handlers when hda changes
-  useEffect(() => {
-    const wsService = wsServiceRef.current;
-    if (!wsService) return;
-    reset();
-    wsService.setHandlers({
-      onStatusChange: (status) => {
-        setWsStatus(status);
-        if (status === "connected") {
-          // Send initial cook request when connected
-          regenerateMesh();
-        }
-      },
-      onStatusLog: (level, log) => {
-        addStatusLog(level, log);
-      },
-      onGeometryData: (data) => {
-        if (data.points && data.indices) {
-          setMeshData({
-            points: data.points,
-            indices: data.indices,
-            normals: data.normals,
-            uvs: data.uvs,
-            colors: data.colors,
-          });
-        }
-      },
-      onFileDownload: (fileName, base64Content) => {
-        SceneTalkConnection.downloadFileFromBase64(fileName, base64Content);
-      },
-      onLatencyUpdate: (latency) => {
-        setLatency(latency);
-      },
-      onRequestComplete: (elapsedTime) => {
-        const elapsedTimeInSeconds = Math.round(elapsedTime);
-        setGenerateTime(elapsedTimeInSeconds.toString());
-
-        flushStatusLog();
-        const currentStatusLog = useSceneStore.getState().statusLog;
-        const errorCount = currentStatusLog.filter(e => e.level === "error").length;
-        const hasErrors = errorCount > 0;
-
-        if (hasErrors) {
-            addStatusLog("error",
-                `Generation completed in ${elapsedTimeInSeconds} ms with errors`);
-        } else {
-            addStatusLog("info",
-                `Generation completed in ${elapsedTimeInSeconds} ms`);
-        }
-        flushStatusLog();
-
-        setRequestInFlight(false);
-
-        if (pendingRequest) {
-            setPendingRequest(false);
-            setTimeout(() => regenerateMesh(), 10);
-        }
-
-        setExportFormat(null);
-      },
-    });
-  }, [selectedHdaId]);
 
   useEffect(() => {
     if (currentWidth > 700 && isModalOpen) {
@@ -152,73 +59,6 @@ export const PackageScene: React.FC = () => {
     }
   }, [currentWidth]);
 
-  // Function to send a regenerate mesh request
-  const regenerateMesh = (format = "raw") => {
-    if (!wsServiceRef.current) {
-      console.error("WebSocket service not initialized");
-      return;
-    }
-
-    if (!selectedHdaId) {
-      console.error("No HDA selected");
-      return;
-    }
-
-    const jd = jobDefinitions?.find(
-      (definition) => definition.source.file_id === selectedHdaId,
-    );
-    if (!jd) {
-      console.error("Failed to find job definition");
-      return;
-    }
-
-    if (requestInFlight) {
-      setPendingRequest(true);
-      return;
-    }
-
-    // Clear the status log
-    clearStatusLog();
-
-    // Set request in flight
-    setRequestInFlight(true);
-    try {
-
-      const dependencyFileIds = jd.params_schema.params['dependencies']?.default || [];
-
-      // Send the cook request with all parameters for the current HDA
-      wsServiceRef.current.sendCookRequestById(
-        selectedHdaId as string,
-        dependencyFileIds as string[],
-        paramValues,
-        inputFiles,
-        format,
-      );
-    } catch (error) {
-      console.error("Error sending cook request:", error);
-      setRequestInFlight(false);
-    }
-  };
-
-  // Watch for export format changes
-  useEffect(() => {
-    if (exportFormat) {
-      regenerateMesh(exportFormat);
-    }
-  }, [exportFormat]);
-
-  // Re-generate mesh when HDA or parameters change
-  useEffect(() => {
-    if (wsStatus === "connected" && !requestInFlight) {
-      regenerateMesh();
-    }
-  }, [paramValues, inputFiles, jobDefinitions]);
-
-  useEffect(() => {
-    if (fileUpload && wsStatus === "connected" && !requestInFlight) {
-      wsServiceRef.current?.sendFileUploadMessage(fileUpload.file, fileUpload.callback);
-    }
-  }, [fileUpload]);
 
   const [jobDef, setJobDef] = useState<JobDefinition | null>(null);
 
@@ -451,7 +291,7 @@ export const PackageScene: React.FC = () => {
             </ModalDialog>
           </Modal>
         )}
-
+        <SceneTalkConnector dependencyFileIds={assetVersion?.contents?.files.map((file) => file.file_id)} />
         <SceneViewer packageName={assetVersion?.name as string} />
         <StatusBar />
       </Box>
