+++ sites/libs/scenetalk/src/index.ts
@@ -1,3 +1,4 @@
 export { SceneTalkConnection } from "./sceneTalkConnection";
 export { useSceneStore } from "./sceneStore";
+export { SceneTalkConnector } from "./sceneTalkConnector";
 export type { InputFile, MeshData, ConnectionStatus } from "./types";
